requirements.txt
pymongo
fpdf
tk
python-dotenv
bcrypt